Pin markdown version to fix build error
markdown made a breaking change between 1.0.0-alpha.20 and 1.0.0-alpha.21
which made genemichaels stop building.

genemichaels never updated (as a library) to fix that,
so if we want to continue using it we have to ping markdown.

The error:

error[E0599]: no variant or associated item named `BlockQuote` found for enum `Node` in the current scope
   --> [path-to]\genemichaels-0.1.21\src\comments.rs:633:15
    |
633 |         Node::BlockQuote(x) => {
    |               ^^^^^^^^^^ variant or associated item not found in `Node`
    |
help: there is a variant with a similar name
    |
633 |         Node::Blockquote(x) => {
    |               ~~~~~~~~~~